some more minor changes

+++ hw1/Player.cpp
@@ -31,7 +31,7 @@ void Player::notifyOnAttackResult(int player, int row, int col, AttackResult res
 
 void Player::setMoves(vector<pair<int, int>> moves)
 {
-    for (int i = 0; i < moves.size(); ++i)
+    for (size_t i = 0; i < moves.size(); ++i)
     {
         //remember moves are from 1 to ROW/COL SIZE while the board is from 0 to ROW/COL SIZE -1
         // we assume that if we got here all the moves are valid
